Be a little bit more resistant to collisions in hashing

+++ synedrion/src/tools/hashing.rs
@@ -32,7 +32,7 @@ pub trait Chain: Sized {
     fn chain_bytes(self, bytes: &(impl AsRef<[u8]> + ?Sized)) -> Self {
         // Hash the length too to prevent hash conflicts. (e.g. H(AB|CD) == H(ABC|D)).
         // Not strictly necessary for fixed-size arrays, but it's easier to just always do it.
-        let len = (bytes.as_ref().len() as u32).to_be_bytes();
+        let len = (bytes.as_ref().len() as u64).to_be_bytes();
         self.chain_raw_bytes(&len).chain_raw_bytes(bytes.as_ref())
     }
 
@@ -46,7 +46,7 @@ pub trait Chain: Sized {
 
     fn chain_slice<T: Hashable>(self, hashable: &[T]) -> Self {
         // Hashing the length too to prevent collisions.
-        let len = hashable.len() as u32;
+        let len = hashable.len() as u64;
         let mut digest = self.chain(&len);
         for elem in hashable {
             digest = digest.chain(elem);
@@ -198,7 +198,7 @@ impl<T: Hashable> Hashable for Vec<T> {
 impl<K: Hashable, V: Hashable> Hashable for BTreeMap<K, V> {
     fn chain<C: Chain>(&self, digest: C) -> C {
         // Hashing the map length too to prevent collisions.
-        let len = self.len() as u32;
+        let len = self.len() as u64;
         let mut digest = digest.chain(&len);
         // The iteration is ordered (by keys)
         for (key, value) in self {
